Moves CachedExecutors, related structs, consts and tests into the program-runtime crate.
